Modify the non-chat function of customer service manager and customer service.Ref#104

Closes #104

See merge request custom-service/HFUT_Group3!124

+++ HFUT_Group3/src/main/java/com/pentaKill/dao/CustomerServiceMapper.java
@@ -41,7 +41,7 @@ public interface CustomerServiceMapper {
 
     public int getCount(int cs_id);
 
-    public int getScore(int cs_id);
+    public double getScore(int cs_id);
 
     public int getTime(int cs_id);
